Remove call to export text to Org
My original thought here is that this step would further santize the
text, but it just ends up removing possibilities that I would expect
to have (such as hard coding in export commands, and keeping
comments). It's extremely easy for the user to export the outputted
complete Org file to Org again (using Org's built in export
dispatcher) if they wish for this extra santization.