Linked Stories additions functionally complete.
Very slow though. All reference updates are now horrifically slow due
to the need to check story language for all files (even when not using
linked stories. This needs to be optimised so that if all stories in a
story pool use the same language, these checks are not carried out for
all files.