Replacing org-next-visible-heading with expected version code
In earier versions of Org, org-next-visible-heading would return nil
if a heading was found, and the end of buffer char position otherwise.
This behaviour has been changed in later versions of Org mode, but Org
Novelist uses the assumption in many places. A simple fix for the
moment is to duplicate the earlier Org mode code as a function within
Org Novelist (org-novelist--next-visble-heading), and call that
instead.

Thanks to @chewxy for spotting this.